+++ src/components/Chips/Chips.tsx
@@ -7,14 +7,16 @@ import { getCSSVar } from "../../services/themes";
 import { elementColorsNames, getElementColor } from "../../utils/colors-vars-map";
 import Avatar from "../Avatar/Avatar";
 import IconButton from "../IconButton/IconButton";
-import { getTestId } from "../../tests/test-ids-utils";
+import Tooltip from "../Tooltip/Tooltip";
+import { ComponentDefaultTestId, getTestId } from "../../tests/test-ids-utils";
 import { ChipsSize } from "./ChipsConstants";
 import { AvatarType } from "../Avatar/AvatarConstants";
 import { SubIcon, VibeComponent, VibeComponentProps } from "../../types";
 import useHover from "../../hooks/useHover/useHover";
 import useSetFocus from "../../hooks/useSetFocus";
-import { ComponentDefaultTestId } from "../../tests/constants";
 import useClickableProps from "../../hooks/useClickableProps/useClickableProps";
+import useIsOverflowing from "../../hooks/useIsOverflowing/useIsOverflowing";
+import useChipOverflowTooltip from "./hooks/useChipOverflowTooltip";
 import { BEMClass } from "../../helpers/bem-helper";
 import "../Clickable/Clickable.scss";
 import styles from "./Chips.module.scss";
@@ -33,8 +35,12 @@ interface ChipsProps extends VibeComponentProps {
   leftIcon?: SubIcon;
   /** Img to place as avatar on the right */
   rightAvatar?: string;
+  /** the type of right avatar */
+  rightAvatarType?: AvatarType;
   /** Img to place as avatar on the left */
   leftAvatar?: string;
+  /** the type of left avatar */
+  leftAvatarType?: AvatarType;
   /** ClassName for left or right icon */
   iconClassName?: string;
   /** ClassName for left or right avatar */
@@ -80,6 +86,7 @@ interface ChipsProps extends VibeComponentProps {
 const Chips: VibeComponent<ChipsProps, HTMLElement> & {
   sizes?: typeof ChipsSize;
   colors?: typeof elementColorsNames;
+  avatarTypes?: typeof AvatarType;
 } = forwardRef<HTMLElement, ChipsProps>(
   (
     {
@@ -100,24 +107,47 @@ const Chips: VibeComponent<ChipsProps, HTMLElement> & {
       onDelete = (_id: string, _e: React.MouseEvent<HTMLSpanElement>) => {},
       onMouseDown,
       onClick,
-      noAnimation = false,
+      noAnimation = true,
       ariaLabel,
       dataTestId,
-      disableClickableBehavior = false
+      disableClickableBehavior = false,
+      leftAvatarType = AvatarType.IMG,
+      rightAvatarType = AvatarType.IMG
     },
     ref
   ) => {
     const overrideDataTestId = dataTestId || getTestId(ComponentDefaultTestId.CHIP, id);
     const hasClickableWrapper = (!!onClick || !!onMouseDown) && !disableClickableBehavior;
     const hasCloseButton = !readOnly && !disabled;
-
-    const focusRef = useRef(null);
-    const [hoverRef, isHovered] = useHover();
-    const { isFocused } = useSetFocus({ ref: focusRef });
+    const overrideAriaLabel = ariaLabel || label;
 
     const iconButtonRef = useRef(null);
+    const labelRef = useRef(null);
     const componentRef = useRef(null);
-    const mergedRef = useMergeRefs({ refs: [ref, componentRef, hoverRef, focusRef] });
+
+    const [hoverRef, isHovered] = useHover();
+    const { isFocused } = useSetFocus({ ref: componentRef });
+    const isOverflowing = useIsOverflowing({ ref: labelRef });
+
+    const mergedRef = useMergeRefs({ refs: [ref, componentRef, hoverRef] });
+
+    const overrideClassName = cx(styles.chips, "chips--wrapper", className, {
+      [styles.disabled]: disabled,
+      [styles.withClose]: hasCloseButton,
+      [styles.noAnimation]: noAnimation,
+      [styles.withUserSelect]: allowTextSelection
+    });
+    const clickableClassName = cx(CLICKABLE_CSS_BASE_CLASS, overrideClassName, {
+      disabled,
+      [clickableBemHelper({ state: "disable-text-selection" })]: !allowTextSelection
+    });
+
+    const overflowProps = useChipOverflowTooltip({
+      isOverflowing,
+      wrapperClassName: overrideClassName,
+      clickableClassName,
+      label
+    });
 
     const backgroundColorStyle = useMemo(() => {
       let cssVar;
@@ -151,17 +181,6 @@ const Chips: VibeComponent<ChipsProps, HTMLElement> & {
       [onClick]
     );
 
-    const overrideClassName = cx(styles.chips, "chips--wrapper", className, {
-      [styles.disabled]: disabled,
-      [styles.withClose]: hasCloseButton,
-      [styles.noAnimation]: noAnimation,
-      [styles.withUserSelect]: allowTextSelection
-    });
-    const clickableClassName = cx(CLICKABLE_CSS_BASE_CLASS, overrideClassName, {
-      disabled,
-      [clickableBemHelper({ state: "disable-text-selection" })]: !allowTextSelection
-    });
-
     const clickableProps = useClickableProps(
       {
         onClick: onClickCallback,
@@ -169,7 +188,7 @@ const Chips: VibeComponent<ChipsProps, HTMLElement> & {
         disabled,
         id,
         dataTestId: overrideDataTestId,
-        ariaLabel: ariaLabel || label,
+        ariaLabel: overrideAriaLabel,
         ariaHidden: false,
         ariaHasPopup: false,
         ariaExpanded: false
@@ -183,72 +202,81 @@ const Chips: VibeComponent<ChipsProps, HTMLElement> & {
           style: backgroundColorStyle
         }
       : {
+          className: overrideClassName,
+          ...overflowProps.wrapperProps,
+          "aria-label": overrideAriaLabel,
+          style: backgroundColorStyle,
           ref: mergedRef,
           onClick: onClickCallback,
           onMouseDown,
           id: id,
-          "data-testid": overrideDataTestId,
-          className: overrideClassName,
-          style: backgroundColorStyle
+          "data-testid": overrideDataTestId
         };
 
+    const leftAvatarProps = leftAvatarType === AvatarType.TEXT ? { text: leftAvatar } : { src: leftAvatar };
+    const rightAvatarProps = leftAvatarType === AvatarType.TEXT ? { text: rightAvatar } : { src: rightAvatar };
+
     return (
-      <div {...wrapperProps}>
-        {leftAvatar ? (
-          <Avatar
-            withoutBorder
-            className={cx(styles.avatar, styles.left, avatarClassName)}
-            customSize={18}
-            src={leftAvatar}
-            type={AvatarType.IMG}
-            key={id}
-          />
-        ) : null}
-        {leftIcon ? (
-          <Icon
-            className={cx(styles.icon, styles.left, iconClassName)}
-            iconType={Icon.type.ICON_FONT}
-            clickable={false}
-            icon={leftIcon}
-            iconSize={iconSize}
-            ignoreFocusStyle
-          />
-        ) : null}
-        <div className={styles.label}>{label}</div>
-        {rightIcon ? (
-          <Icon
-            className={cx(styles.icon, styles.right, iconClassName)}
-            iconType={Icon.type.ICON_FONT}
-            clickable={false}
-            icon={rightIcon}
-            iconSize={iconSize}
-            ignoreFocusStyle
-          />
-        ) : null}
-        {rightAvatar ? (
-          <Avatar
-            withoutBorder
-            className={cx(styles.avatar, styles.right, avatarClassName)}
-            customSize={16}
-            src={rightAvatar}
-            type={AvatarType.IMG}
-            key={id}
-          />
-        ) : null}
-        {hasCloseButton && (
-          <IconButton
-            size={ChipsSize.XXS}
-            color={IconButton.colors.ON_PRIMARY_COLOR}
-            className={cx(styles.icon, styles.close)}
-            ariaLabel="Remove"
-            hideTooltip
-            icon={CloseSmall}
-            onClick={onDeleteCallback}
-            dataTestId={`${overrideDataTestId}-close`}
-            ref={iconButtonRef}
-          />
-        )}
-      </div>
+      <Tooltip {...overflowProps.tooltipProps}>
+        <div {...wrapperProps}>
+          {leftAvatar ? (
+            <Avatar
+              withoutBorder
+              className={cx(styles.avatar, styles.left, avatarClassName)}
+              customSize={18}
+              type={leftAvatarType}
+              key={id}
+              {...leftAvatarProps}
+            />
+          ) : null}
+          {leftIcon ? (
+            <Icon
+              className={cx(styles.icon, styles.left, iconClassName)}
+              iconType={Icon.type.ICON_FONT}
+              clickable={false}
+              icon={leftIcon}
+              iconSize={iconSize}
+              ignoreFocusStyle
+            />
+          ) : null}
+          <div className={styles.label} ref={labelRef}>
+            {label}
+          </div>
+          {rightIcon ? (
+            <Icon
+              className={cx(styles.icon, styles.right, iconClassName)}
+              iconType={Icon.type.ICON_FONT}
+              clickable={false}
+              icon={rightIcon}
+              iconSize={iconSize}
+              ignoreFocusStyle
+            />
+          ) : null}
+          {rightAvatar ? (
+            <Avatar
+              withoutBorder
+              className={cx(styles.avatar, styles.right, avatarClassName)}
+              customSize={16}
+              type={rightAvatarType}
+              key={id}
+              {...rightAvatarProps}
+            />
+          ) : null}
+          {hasCloseButton && (
+            <IconButton
+              size={ChipsSize.XXS}
+              color={IconButton.colors.ON_PRIMARY_COLOR}
+              className={cx(styles.icon, styles.close)}
+              ariaLabel="Remove"
+              hideTooltip
+              icon={CloseSmall}
+              onClick={onDeleteCallback}
+              dataTestId={`${overrideDataTestId}-close`}
+              ref={iconButtonRef}
+            />
+          )}
+        </div>
+      </Tooltip>
     );
   }
 );
@@ -256,7 +284,8 @@ const Chips: VibeComponent<ChipsProps, HTMLElement> & {
 Object.assign(Chips, {
   sizes: ChipsSize,
   defaultTestId: ComponentDefaultTestId.CHIP,
-  colors: elementColorsNames
+  colors: elementColorsNames,
+  avatarTypes: AvatarType
 });
 
 export default Chips;