+++ src/components/Dropdown/Dropdown.jsx
@@ -23,6 +23,7 @@ const Dropdown = ({
   className,
   placeholder,
   disabled,
+  readOnly,
   onMenuOpen,
   onMenuClose,
   onFocus,
@@ -68,7 +69,10 @@ const Dropdown = ({
   insideOverflowWithTransformContainer,
   ref,
   tooltipContent,
-  onKeyDown
+  onKeyDown,
+  isLoading,
+  loadingMessage,
+  ariaLabel
 }) => {
   const controlRef = useRef();
   const overrideDefaultValue = useMemo(() => {
@@ -94,6 +98,14 @@ const Dropdown = ({
     return {};
   }, [selectedOptions]);
 
+  const overrideAriaLabel = useMemo(() => {
+    return (
+      ariaLabel ||
+      `${readOnly ? "Readonly " : ""} ${tooltipContent} ${
+        Array.isArray(selectedOptions) ? `Selected: ${selectedOptions.map(o => o.label).join(", ")}` : "Select"
+      }`
+    );
+  }, [ariaLabel, readOnly, selectedOptions, tooltipContent]);
   const value = multi ? selectedOptions : customValue;
 
   const styles = useMemo(() => {
@@ -154,8 +166,15 @@ const Dropdown = ({
   const Input = useCallback(props => <components.Input {...props} aria-label="Dropdown input" />, []);
 
   const SingleValue = useCallback(
-    props => <SingleValueComponent {...props} Renderer={finalValueRenderer} selectedOption={selectedOptions[0]} />,
-    [finalValueRenderer, selectedOptions]
+    props => (
+      <SingleValueComponent
+        {...props}
+        readOnly={readOnly}
+        Renderer={finalValueRenderer}
+        selectedOption={selectedOptions[0]}
+      />
+    ),
+    [finalValueRenderer, readOnly, selectedOptions]
   );
 
   const ClearIndicator = useCallback(props => <ClearIndicatorComponent {...props} size={size} />, [size]);
@@ -276,8 +295,12 @@ const Dropdown = ({
       noOptionsMessage={noOptionsMessage}
       placeholder={placeholder}
       isDisabled={disabled}
-      isClearable={clearable}
-      isSearchable={searchable}
+      isClearable={!readOnly && clearable}
+      isSearchable={!readOnly && searchable}
+      readOnly={readOnly}
+      aria-readonly={readOnly}
+      aria-label={overrideAriaLabel}
+      aria-details={tooltipContent}
       defaultValue={defaultValue}
       value={value}
       onMenuOpen={onMenuOpen}
@@ -295,7 +318,7 @@ const Dropdown = ({
       maxMenuHeight={maxMenuHeight}
       menuPortalTarget={menuPortalTarget}
       menuPlacement={menuPlacement}
-      menuIsOpen={menuIsOpen}
+      menuIsOpen={!readOnly && menuIsOpen}
       tabIndex={tabIndex}
       id={id}
       autoFocus={autoFocus}
@@ -303,7 +326,8 @@ const Dropdown = ({
       ref={ref}
       withMandatoryDefaultOptions={withMandatoryDefaultOptions}
       isOptionSelected={isOptionSelected}
-      aria-details={tooltipContent}
+      isLoading={isLoading}
+      loadingMessage={loadingMessage}
       {...asyncAdditions}
       {...additions}
     />
@@ -339,7 +363,12 @@ Dropdown.defaultProps = {
   withMandatoryDefaultOptions: false,
   insideOverflowContainer: false,
   insideOverflowWithTransformContainer: false,
-  tooltipContent: ""
+  tooltipContent: "",
+  disabled: false,
+  readOnly: false,
+  isLoading: false,
+  loadingMessage: undefined,
+  ariaLabel: undefined
 };
 
 Dropdown.propTypes = {
@@ -355,6 +384,10 @@ Dropdown.propTypes = {
    * If set to true, dropdown will be disabled
    */
   disabled: PropTypes.bool,
+  /**
+   * If set to true, dropdown won't be editable
+   */
+  readOnly: PropTypes.bool,
   /**
    * Called when menu is opened
    */
@@ -549,7 +582,19 @@ Dropdown.propTypes = {
   /**
    * When content is passed, the dropdown will include a tooltip on the dropdown's value.
    */
-  tooltipContent: PropTypes.string
+  tooltipContent: PropTypes.string,
+  /**
+   * Display the drop down with loading state.
+   */
+  isLoading: PropTypes.bool,
+  /**
+   * Overrides the built-in logic of loading message design
+   */
+  loadingMessage: PropTypes.func,
+  /**
+   * aria-label attribute for dropdown
+   */
+  ariaLabel: PropTypes.string
 };
 
 export default Dropdown;