+++ src/components/Dropdown/components/ChildrenContent/ChildrenContent.jsx
@@ -1,10 +1,20 @@
 import Avatar from "../../../Avatar/Avatar";
 import Icon from "../../../Icon/Icon";
+import { useCallback } from "react";
 import styles from "./ChildrenContent.module.scss";
 
-export const ChildrenContent = ({ data, children }) => {
+export const ChildrenContent = ({ data, children, readOnly }) => {
+  const onMouseDown = useCallback(
+    event => {
+      if (readOnly) {
+        event.stopPropagation();
+      }
+    },
+    [readOnly]
+  );
+
   return (
-    <div className={styles.valueContainer}>
+    <div className={styles.valueContainer} onMouseDown={onMouseDown}>
       {data?.leftAvatar && (
         <Avatar withoutBorder square={data.square} src={data.leftAvatar} type={Avatar.types.IMG} customSize={18} />
       )}