+++ src/components/EditableHeading/EditableHeading.tsx
@@ -1,4 +1,3 @@
-/* eslint-disable react/destructuring-assignment */
 import React, { useCallback, useEffect, useLayoutEffect, useRef, useState } from "react";
 import cx from "classnames";
 import Heading, { HeadingProps } from "../Heading/Heading";
@@ -8,6 +7,7 @@ import usePrevious from "../../hooks/usePrevious";
 import { InputType } from "../EditableInput/EditableInputConstants";
 import "./EditableHeading.scss";
 import { HeadingSizes, HeadingTypes } from "../Heading/HeadingConstants";
+import { Sizes } from "../../constants";
 
 export interface EditableHeadingProps extends EditableInputProps, HeadingProps {
   displayPlaceholderInTextMode?: boolean;
@@ -27,7 +27,7 @@ export interface EditableHeadingProps extends EditableInputProps, HeadingProps {
 }
 
 const EditableHeading: React.FC<EditableHeadingProps> & {
-  sizes?: typeof HeadingSizes;
+  sizes?: Sizes;
   types?: typeof HeadingTypes;
 } = props => {
   const {
@@ -150,7 +150,7 @@ const EditableHeading: React.FC<EditableHeadingProps> & {
       tooltipPosition: props.tooltipPosition,
       ellipsisMaxLines: props.ellipsisMaxLines,
       nonEllipsisTooltip: props.tooltip,
-      size: size as typeof HeadingSizes,
+      size: size as Sizes,
       highlightTerm
     };