+++ src/components/Tabs/Tab/Tab.tsx
@@ -1,14 +1,22 @@
 /* eslint-disable jsx-a11y/anchor-is-valid,jsx-a11y/click-events-have-key-events */
+import cx from "classnames";
 import React, { FC, forwardRef, ReactElement, useRef } from "react";
 import { noop as NOOP } from "lodash-es";
-import cx from "classnames";
 import useMergeRefs from "../../../hooks/useMergeRefs";
 import Icon, { IconSubComponentProps } from "../../Icon/Icon";
 import VibeComponentProps from "../../../types/VibeComponentProps";
 import { IconType } from "../../Icon/IconConstants";
-import "./Tab.scss";
+import { ComponentDefaultTestId, getTestId } from "../../../tests/test-ids-utils";
+import styles from "./Tab.module.scss";
 
 export interface TabProps extends VibeComponentProps {
+  /**
+   * Class name for tab link-name
+   */
+  tabInnerClassName?: string;
+  /**
+   * Tab index
+   */
   value?: number;
   disabled?: boolean;
   active?: boolean;
@@ -17,6 +25,9 @@ export interface TabProps extends VibeComponentProps {
   iconType?: IconType;
   iconSide?: string;
   onClick?: (value: number) => void;
+  /**
+   * Tab link-name
+   */
   children?: string | ReactElement[];
 }
 
@@ -24,6 +35,7 @@ const Tab: FC<TabProps> = forwardRef(
   (
     {
       className,
+      tabInnerClassName,
       id,
       value = 0,
       disabled = false,
@@ -33,7 +45,8 @@ const Tab: FC<TabProps> = forwardRef(
       icon,
       iconType,
       iconSide = "left",
-      children
+      children,
+      "data-testid": dataTestId
     },
     ref
   ) => {
@@ -49,7 +62,8 @@ const Tab: FC<TabProps> = forwardRef(
           ariaHidden={true}
           iconType={iconType}
           icon={icon}
-          className={cx("tab-icon", iconSide)}
+          // css-modules-migration: iconSide - used only for overrides
+          className={cx(styles.tabIcon, "tab-icon", iconSide)}
           iconSize={18}
           ignoreFocusStyle
         />
@@ -65,17 +79,21 @@ const Tab: FC<TabProps> = forwardRef(
       <li
         ref={mergedRef}
         key={id}
-        className={cx("tab--wrapper", className, {
-          active,
-          disabled,
-          "tab-focus-visible-inset": focus
+        className={cx(styles.tabWrapper, "tab--wrapper", className, {
+          [styles.active]: active,
+          ["active"]: active,
+          [styles.disabled]: disabled,
+          ["disabled"]: disabled,
+          [styles.tabFocusVisibleInset]: focus,
+          ["tab-focus-visible-inset"]: focus
         })}
         id={id}
         role="tab"
         aria-selected={active}
         aria-disabled={disabled}
+        data-testid={dataTestId || getTestId(ComponentDefaultTestId.TAB, id)}
       >
-        <a className="tab-inner" onClick={() => !disabled && onClick(value)}>
+        <a className={cx(styles.tabInner, "tab-inner", tabInnerClassName)} onClick={() => !disabled && onClick(value)}>
           {renderIconAndChildren()}
         </a>
       </li>