+++ src/components/Tooltip/Tooltip.tsx
@@ -1,8 +1,8 @@
 /* eslint-disable react/jsx-props-no-spreading */
+import { camelCase, isFunction } from "lodash-es";
+import cx from "classnames";
 import React, { CSSProperties, isValidElement, PureComponent, ReactElement } from "react";
-import classnames from "classnames";
 import { Modifier } from "react-popper";
-import { isFunction } from "lodash-es";
 import Dialog from "../Dialog/Dialog";
 import { AnimationType, BASE_SIZES_WITH_NONE, HideShowEvent, JustifyType } from "../../constants";
 import { DialogPosition } from "../../constants/positions";
@@ -10,7 +10,8 @@ import VibeComponentProps from "../../types/VibeComponentProps";
 import { TooltipArrowPosition, TooltipTheme } from "./TooltipConstants";
 import { ElementContent } from "../../types/ElementContent";
 import { MoveBy } from "../../types/MoveBy";
-import "./Tooltip.scss";
+import styles from "./Tooltip.module.scss";
+import { getStyle } from "../../helpers/typesciptCssModulesHelper";
 
 // TODO TS-migration extend DialogProps, once Dialog is migrated to TS
 export interface TooltipProps extends VibeComponentProps {
@@ -76,11 +77,22 @@ export interface TooltipProps extends VibeComponentProps {
    * an array of hide/show trigger - Tooltip.hideShowTriggers
    */
   hideTrigger?: HideShowEvent | Array<HideShowEvent>;
+  /**
+   * If true, prevents open Tooltip from closing on mouseEnter and closes Tooltip, when mouse leaves it
+   */
   showOnDialogEnter?: boolean;
   /**
    * A Classname to be added to <spam> element which wraps the children
    */
   referenceWrapperClassName?: string;
+  /**
+   * Treats keyboard focus/blur events as mouse-enter/mouse-leave events
+   */
+  addKeyboardHideShowTriggersByDefault?: boolean;
+  /**
+   * set the state of the tooltip - open/close - controlled component
+   */
+  open?: boolean;
 }
 // When last tooltip was shown in the last 1.5 second - the next tooltip will be shown immediately
 const IMMEDIATE_SHOW_THRESHOLD_MS = 1500;
@@ -117,7 +129,9 @@ export default class Tooltip extends PureComponent<TooltipProps> {
     showTrigger: Tooltip.hideShowTriggers.MOUSE_ENTER,
     hideTrigger: Tooltip.hideShowTriggers.MOUSE_LEAVE,
     showOnDialogEnter: false,
-    referenceWrapperClassName: ""
+    referenceWrapperClassName: "",
+    addKeyboardHideShowTriggersByDefault: false,
+    open: false
   };
   constructor(props: TooltipProps) {
     super(props);
@@ -154,8 +168,13 @@ export default class Tooltip extends PureComponent<TooltipProps> {
     return (
       <div
         style={style}
-        className={classnames(
-          `monday-style-tooltip monday-style-tooltip-${theme} padding-size-${paddingSize}`,
+        className={cx(
+          styles.tooltip,
+          "monday-style-tooltip",
+          getStyle(styles, camelCase(theme)),
+          `monday-style-tooltip-${theme}`,
+          getStyle(styles, camelCase("padding-size-" + paddingSize)),
+          `padding-size-${paddingSize}`,
           className
         )}
       >
@@ -218,7 +237,9 @@ export default class Tooltip extends PureComponent<TooltipProps> {
       tip,
       showTrigger,
       hideTrigger,
-      showOnDialogEnter
+      showOnDialogEnter,
+      addKeyboardHideShowTriggersByDefault,
+      open
     } = this.props;
 
     if (!children) {
@@ -231,19 +252,28 @@ export default class Tooltip extends PureComponent<TooltipProps> {
     const content = this.renderTooltipContent;
     const dialogProps = {
       ...this.props,
+      open,
       startingEdge: justify,
       tooltip: tip,
       content,
       getContainer: getContainer || this.getContainer,
       moveBy,
-      tooltipClassName: `monday-style-arrow monday-style-arrow-${theme} padding-size-${paddingSize}`,
+      tooltipClassName: cx(
+        styles.arrow,
+        "monday-style-arrow",
+        getStyle(styles, theme),
+        `monday-style-arrow-${theme}`,
+        getStyle(styles, camelCase("padding-size-" + paddingSize)),
+        `padding-size-${paddingSize}`
+      ),
       animationType: AnimationType.EXPAND,
       onDialogDidHide: this.onTooltipHide,
       onDialogDidShow: this.onTooltipShow,
       getDynamicShowDelay: this.getShowDelay,
       showTrigger,
       hideTrigger,
-      showOnDialogEnter
+      showOnDialogEnter,
+      addKeyboardHideShowTriggersByDefault
     };
     return <Dialog {...dialogProps}>{children}</Dialog>;
   }