+++ src/helpers/typesciptCssModulesHelper.ts
@@ -3,7 +3,10 @@
  * @param styles modular styles object
  * @param key string classname
  */
-export function getStyle<StylesType>(styles: StylesType, key: string | undefined | null) {
-  if (!key) return "";
+export function getStyle<StylesType>(
+  styles: StylesType,
+  key: string | undefined | null
+): StylesType[keyof StylesType] | string {
+  if (!key || !styles[key as keyof typeof styles]) return "";
   return styles[key as keyof typeof styles];
 }